Use Factory to create ArtifactClient
