+++ packages/artifact/__tests__/upload.test.ts
@@ -32,7 +32,7 @@ jest.mock('@actions/http-client')
 describe('Upload Tests', () => {
   beforeAll(async () => {
     // mock all output so that there is less noise when running tests
-    console.log = jest.fn()
+    jest.spyOn(console, 'log').mockImplementation(() => {})
     jest.spyOn(core, 'debug').mockImplementation(() => {})
     jest.spyOn(core, 'info').mockImplementation(() => {})
     jest.spyOn(core, 'warning').mockImplementation(() => {})
