+++ packages/artifact/__tests__/util.test.ts
@@ -6,10 +6,9 @@ import {getRuntimeUrl, getWorkFlowRunId} from '../src/internal-config-variables'
 jest.mock('../src/internal-config-variables')
 
 describe('Utils', () => {
-
   beforeAll(() => {
     // mock all output so that there is less noise when running tests
-    console.log = jest.fn()
+    jest.spyOn(console, 'log').mockImplementation(() => {})
     jest.spyOn(core, 'debug').mockImplementation(() => {})
     jest.spyOn(core, 'info').mockImplementation(() => {})
     jest.spyOn(core, 'warning').mockImplementation(() => {})
