+++ packages/artifact/src/artifact-client.ts
@@ -1,0 +1,9 @@
+import {ArtifactClient, DefaultArtifactClient} from './internal-artifact-client'
+export {ArtifactClient}
+
+/**
+ * Constructs an ArtifactClient
+ */
+export function create(): ArtifactClient {
+  return DefaultArtifactClient.create()
+}
